requirements.txt
Flask
pymongo[srv]
python-dotenv
Flask-WTF
flask_pymongo